New progress bar logic
Added progress bar logic
Added error messages *basic
Added color effects backgrounds

+++ views/posts/index.php
@@ -17,16 +17,19 @@
         <td><?=cutLongText($post['content'])?></td>
         <td><?=htmlspecialchars($post['date'])?></td>
         <td><?=htmlspecialchars($post['full_name'])?></td>
-
+    <?php if ($username = htmlspecialchars($_SESSION['username']) == 'admin') { ?>
         <td>
             <a href="<?=APP_ROOT?>/posts/edit/<?=
             htmlspecialchars($post['id'])?>">[Edit]</a>
             <a href="<?=APP_ROOT?>/posts/delete/<?=
             htmlspecialchars($post['id'])?>">[Delete]</a>
         </td>
-
+    <?php } else {?>
+        <td>
+            <p>Administrator</p>
+        </td>
+        <?php } ?>
     </tr>
-
 <?php endforeach ?>
     </table>
 </main>
